+++ packages/vendure-plugin-frequently-bought-together/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.2.5 (2026-02-03)
+
+- Updated license to MIT
+
+# 1.2.4 (2026-02-03)
+
+- Documentation update: remove paid plugin notice, free for all!
+
 # 1.2.3 (2025-11-13)
 
 - Documentation update